+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Upcoming
 
 - refactor: add `WithState` class to be the return value of the `store.with_state` so that it can report correct signature of its `__call__` method
+- chore(docs): add documentation strings for elements defined in `autorun.py`
 
 ## Version 0.20.2